add barelycomposer api
