+++ barelymusician/barelymusician.h
@@ -268,7 +268,7 @@ enum BarelyOscillatorShape_Values {
   BarelyOscillatorShape_kSquare,
   /// White noise.
   BarelyOscillatorShape_kNoise,
-  /// Count.
+  /// Number of oscillator shapes.
   BarelyOscillatorShape_kCount,
 };
 
@@ -285,7 +285,7 @@ enum BarelySamplePlaybackMode_Values {
   BarelySamplePlaybackMode_kSustain,
   /// Loop.
   BarelySamplePlaybackMode_kLoop,
-  /// Count.
+  /// Number of playback modes.
   BarelySamplePlaybackMode_kCount,
 };
 
@@ -303,21 +303,6 @@ typedef struct BarelySampleDataSlice {
   /// Number of mono samples.
   int32_t sample_count;
 } BarelySampleDataSlice;
-
-/// A musical scale.
-typedef struct BarelyScale {
-  /// Array of note pitches relative to the root note pitch.
-  const double* pitches;
-
-  /// Number of note pitches.
-  int32_t pitch_count;
-
-  /// Root note pitch of the scale.
-  double root_pitch;
-
-  /// Mode index.
-  int32_t mode;
-} BarelyScale;
 
 /// Note off event create callback signature.
 ///
@@ -741,15 +726,6 @@ BARELY_EXPORT bool BarelyPerformer_Start(BarelyPerformerHandle performer);
 /// @return True if successful, false otherwise.
 BARELY_EXPORT bool BarelyPerformer_Stop(BarelyPerformerHandle performer);
 
-/// Gets a scale note pitch for a given degree.
-///
-/// @param scale Pointer to scale.
-/// @param degree Scale degree.
-/// @param out_pitch Output note pitch.
-/// @return True if successful, false otherwise.
-BARELY_EXPORT bool BarelyScale_GetPitch(const BarelyScale* scale, int32_t degree,
-                                        double* out_pitch);
-
 /// Gets the position of a task.
 ///
 /// @param task Task handle.
@@ -827,7 +803,7 @@ enum class OscillatorShape : BarelyOscillatorShape {
   kSquare = BarelyOscillatorShape_kSquare,
   /// White noise.
   kNoise = BarelyOscillatorShape_kNoise,
-  /// Count.
+  /// Number of oscillator shapes.
   kCount = BarelyOscillatorShape_kCount,
 };
 
@@ -841,7 +817,7 @@ enum class SamplePlaybackMode : BarelySamplePlaybackMode {
   kSustain = BarelySamplePlaybackMode_kSustain,
   /// Loop.
   kLoop = BarelySamplePlaybackMode_kLoop,
-  /// Count.
+  /// Number of playback modes.
   kCount = BarelySamplePlaybackMode_kCount,
 };
 
@@ -865,50 +841,6 @@ struct SampleDataSlice : public BarelySampleDataSlice {
   // NOLINTNEXTLINE(google-explicit-constructor)
   constexpr SampleDataSlice(BarelySampleDataSlice sample_data_slice) noexcept
       : BarelySampleDataSlice{sample_data_slice} {}
-};
-
-/// A musical scale.
-struct Scale : public BarelyScale {
- public:
-  /// Default constructor.
-  constexpr Scale() noexcept = default;
-
-  /// Constructs a new `Scale`.
-  ///
-  /// @param pitches Span of pitches.
-  /// @param root_pitch Root pitch.
-  /// @param mode Mode.
-  constexpr Scale(std::span<const double> pitches, double root_pitch = 0.0, int mode = 0) noexcept
-      : Scale(BarelyScale{pitches.data(), static_cast<int32_t>(pitches.size()), root_pitch,
-                          static_cast<int32_t>(mode)}) {}
-
-  /// Constructs a new `Scale` from a raw type.
-  ///
-  /// @param scale Raw scale.
-  // NOLINTNEXTLINE(google-explicit-constructor)
-  constexpr Scale(BarelyScale scale) noexcept : BarelyScale{scale} {
-    assert(scale.pitches != nullptr);
-    assert(scale.pitch_count > 0);
-    assert(scale.mode >= 0 && scale.mode < scale.pitch_count);
-  }
-
-  /// Returns the pitch for a given degree.
-  ///
-  /// @param degree Degree.
-  /// @return Pitch.
-  [[nodiscard]] double GetPitch(int degree) const noexcept {
-    double pitch = 0.0;
-    [[maybe_unused]] const bool success = BarelyScale_GetPitch(this, degree, &pitch);
-    assert(success);
-    return pitch;
-  }
-
-  /// Returns the number of pitches in the scale.
-  ///
-  /// @return Number of pitches.
-  [[nodiscard]] constexpr int GetPitchCount() const noexcept {
-    return static_cast<int>(pitch_count);
-  }
 };
 
 /// Note off event.
@@ -1507,7 +1439,7 @@ class MusicianHandle : public HandleWrapper<BarelyMusicianHandle> {
 
   /// Destroys a `MusicianHandle`.
   ///
-  /// @param musician Musician.
+  /// @param musician Musician handle.
   static void Destroy(MusicianHandle musician) noexcept { BarelyMusician_Destroy(musician); }
 
   /// Default constructor.
