+++ barelymusician/barelymusician_test.cpp
@@ -1,7 +1,5 @@
 #include "barelymusician/barelymusician.h"
 
-#include <array>
-
 #include "gtest/gtest.h"
 
 namespace barely {
@@ -64,26 +62,6 @@ TEST(MusicianTest, CreateDestroyPerformer) {
   Musician musician(1);
   auto performer = musician.AddPerformer();
   musician.RemovePerformer(performer);
-}
-
-TEST(ScaleTest, GetPitch) {
-  constexpr int kPitchCount = 5;
-  const std::array<double, kPitchCount> kPitches = {0.0, 0.2, 0.35, 0.5, 0.95};
-  constexpr double kRootPitch = 1.75;
-  constexpr int kMode = 1;
-
-  const Scale scale = {kPitches, kRootPitch, kMode};
-
-  constexpr int kOctaveRange = 2;
-  for (int octave = -kOctaveRange; octave <= kOctaveRange; ++octave) {
-    for (int i = 0; i < kPitchCount; ++i) {
-      const int degree = octave * kPitchCount + i;
-      // NOLINTNEXTLINE(bugprone-integer-division)
-      const double expected_pitch = kRootPitch + octave + (i + kMode) / kPitchCount +
-                                    kPitches[(i + kMode) % kPitchCount] - kPitches[kMode];
-      EXPECT_EQ(scale.GetPitch(degree), expected_pitch) << degree;
-    }
-  }
 }
 
 }  // namespace
