+++ barelymusician/components/CMakeLists.txt
@@ -5,8 +5,12 @@ add_library(
 )
 target_link_libraries(
   barelymusician_components_arpeggiator
-  barelymusician
+  barelycomposer_headeronly
+  barelymusician_headeronly
   barelymusician_common_random
+  barelymusician_internal_instrument_controller
+  barelymusician_internal_musician
+  barelymusician_internal_performer
 )
 
 add_library(
@@ -16,6 +20,10 @@ add_library(
 )
 target_link_libraries(
   barelymusician_components_repeater
-  barelymusician
+  barelycomposer_headeronly
+  barelymusician_headeronly
   barelymusician_common_random
+  barelymusician_internal_instrument_controller
+  barelymusician_internal_musician
+  barelymusician_internal_performer
 )
