+++ barelymusician/components/repeater.cpp
@@ -1,134 +1,48 @@
 #include "barelymusician/components/repeater.h"
 
-#include <cstdint>
 #include <optional>
 #include <utility>
 
+#include "barelymusician/barelycomposer.h"
 #include "barelymusician/barelymusician.h"
+#include "barelymusician/internal/instrument_controller.h"
+#include "barelymusician/internal/musician.h"
 
-// Repeater.
-struct BarelyRepeater : public barely::Repeater {
-  // Constructs `BarelyRepeater` with `musician` and `process_order`.
-  BarelyRepeater(BarelyMusicianHandle musician, int process_order) noexcept
-      : barely::Repeater(barely::MusicianHandle(musician), process_order) {}
+namespace barely::internal {
 
-  // Destroys `BarelyRepeater`.
-  ~BarelyRepeater() noexcept { SetInstrument(std::nullopt); }
-};
+namespace {
 
-bool BarelyRepeater_Clear(BarelyRepeaterHandle repeater) {
-  if (!repeater) return false;
+constexpr double kNoteIntensity = 1.0;
 
-  repeater->Clear();
-  return true;
-}
-
-bool BarelyRepeater_Create(BarelyMusician* musician, int32_t process_order,
-                           BarelyRepeaterHandle* out_repeater) {
-  if (!musician || !out_repeater) return false;
-
-  *out_repeater = new BarelyRepeater(musician, static_cast<int>(process_order));
-  return true;
-}
-
-bool BarelyRepeater_Destroy(BarelyRepeaterHandle repeater) {
-  if (!repeater) return false;
-
-  delete repeater;
-  return true;
-}
-
-bool BarelyRepeater_IsPlaying(BarelyRepeaterHandle repeater, bool* out_is_playing) {
-  if (!repeater || !out_is_playing) return false;
-
-  *out_is_playing = repeater->IsPlaying();
-  return true;
-}
-
-bool BarelyRepeater_Pop(BarelyRepeaterHandle repeater) {
-  if (!repeater) return false;
-
-  repeater->Pop();
-  return true;
-}
-
-bool BarelyRepeater_Push(BarelyRepeaterHandle repeater, double pitch, int32_t length) {
-  if (!repeater) return false;
-
-  repeater->Push(pitch, static_cast<int>(length));
-  return true;
-}
-
-bool BarelyRepeater_PushSilence(BarelyRepeaterHandle repeater, int32_t length) {
-  if (!repeater) return false;
-
-  repeater->Push(std::nullopt, static_cast<int>(length));
-  return true;
-}
-
-bool BarelyRepeater_SetInstrument(BarelyRepeaterHandle repeater,
-                                  BarelyInstrumentHandle instrument) {
-  if (!repeater) return false;
-
-  repeater->SetInstrument(
-      instrument != nullptr ? std::optional(barely::InstrumentHandle(instrument)) : std::nullopt);
-  return true;
-}
-
-bool BarelyRepeater_SetRate(BarelyRepeaterHandle repeater, double rate) {
-  if (!repeater) return false;
-
-  repeater->SetRate(rate);
-  return true;
-}
-
-bool BarelyRepeater_SetStyle(BarelyRepeaterHandle repeater, BarelyRepeaterStyle style) {
-  if (!repeater) return false;
-
-  repeater->SetStyle(static_cast<barely::RepeaterStyle>(style));
-  return true;
-}
-
-bool BarelyRepeater_Start(BarelyRepeaterHandle repeater, double pitch_offset) {
-  if (!repeater) return false;
-
-  repeater->Start(pitch_offset);
-  return true;
-}
-
-bool BarelyRepeater_Stop(BarelyRepeaterHandle repeater) {
-  if (!repeater) return false;
-
-  repeater->Stop();
-  return true;
-}
-
-namespace barely {
+}  // namespace
 
 // NOLINTNEXTLINE(bugprone-exception-escape)
-Repeater::Repeater(MusicianHandle musician, int process_order) noexcept
-    : musician_(std::move(musician)), performer_(musician_.AddPerformer(process_order)) {
-  performer_.SetLooping(true);
-  performer_.SetLoopLength(1.0);
-  performer_.AddTask(
-      [this]() noexcept {
-        if (pitches_.empty() || !Update() || !instrument_.has_value()) {
-          return;
-        }
-        const auto& [pitch_or, length] = pitches_[index_];
-        if (!pitches_[index_].first.has_value()) {
-          return;
-        }
-        const double pitch = *pitches_[index_].first + pitch_offset_;
-        instrument_->SetNoteOn(pitch);
-        performer_.ScheduleOneOffTask([this, pitch]() { instrument_->SetNoteOff(pitch); },
-                                      static_cast<double>(length) * performer_.GetLoopLength());
-      },
-      0.0);
+Repeater::Repeater(Musician& musician, int process_order) noexcept
+    : musician_(musician), performer_(musician_.AddPerformer(process_order)) {
+  performer_->SetLooping(true);
+  performer_->SetLoopLength(1.0);
+  // TODO(#126): This should not need `TaskEvent::Callback`.
+  TaskEvent::Callback callback = [this]() noexcept {
+    if (pitches_.empty() || !Update() || instrument_ == nullptr) {
+      return;
+    }
+    const auto& [pitch_or, length] = pitches_[index_];
+    if (!pitches_[index_].first.has_value()) {
+      return;
+    }
+    const double pitch = *pitches_[index_].first + pitch_offset_;
+    instrument_->SetNoteOn(pitch, kNoteIntensity);
+    TaskEvent::Callback note_off_callback = [this, pitch]() noexcept {
+      instrument_->SetNoteOff(pitch);
+    };
+    performer_->ScheduleOneOffTask(EventWithCallback<TaskEvent>(note_off_callback),
+                                   static_cast<double>(length) * performer_->GetLoopLength());
+  };
+  performer_->AddTask(EventWithCallback<TaskEvent>(callback), 0.0);
 }
 
 Repeater::~Repeater() noexcept {
-  if (IsPlaying() && instrument_.has_value()) {
+  if (IsPlaying() && instrument_ != nullptr) {
     instrument_->SetAllNotesOff();
   }
   musician_.RemovePerformer(performer_);
@@ -136,7 +50,7 @@ Repeater::~Repeater() noexcept {
 
 void Repeater::Clear() noexcept { pitches_.clear(); }
 
-bool Repeater::IsPlaying() const noexcept { return performer_.IsPlaying(); }
+bool Repeater::IsPlaying() const noexcept { return performer_->IsPlaying(); }
 
 // NOLINTNEXTLINE(bugprone-exception-escape)
 void Repeater::Pop() noexcept {
@@ -144,7 +58,7 @@ void Repeater::Pop() noexcept {
     return;
   }
   if (index_ == static_cast<int>(pitches_.size()) - 1 && IsPlaying()) {
-    if (instrument_.has_value()) {
+    if (instrument_ != nullptr) {
       instrument_->SetNoteOff(*pitches_.back().first + pitch_offset_);
     }
     remaining_length_ = 0;
@@ -157,8 +71,8 @@ void Repeater::Push(std::optional<double> pitch_or, int length) noexcept {
   pitches_.emplace_back(pitch_or, length);
 }
 
-void Repeater::SetInstrument(std::optional<InstrumentHandle> instrument) noexcept {
-  if (instrument_.has_value() && IsPlaying()) {
+void Repeater::SetInstrument(InstrumentController* instrument) noexcept {
+  if (instrument_ != nullptr) {
     instrument_->SetAllNotesOff();
   }
   instrument_ = instrument;
@@ -166,7 +80,7 @@ void Repeater::SetInstrument(std::optional<InstrumentHandle> instrument) noexcep
 
 void Repeater::SetRate(double rate) noexcept {
   const double length = (rate > 0.0) ? 1.0 / rate : 0.0;
-  performer_.SetLoopLength(length);
+  performer_->SetLoopLength(length);
 }
 
 void Repeater::SetStyle(RepeaterStyle style) noexcept { style_ = style; }
@@ -177,7 +91,7 @@ void Repeater::Start(double pitch_offset) noexcept {
     return;
   }
   pitch_offset_ = pitch_offset;
-  performer_.Start();
+  performer_->Start();
 }
 
 // NOLINTNEXTLINE(bugprone-exception-escape)
@@ -185,9 +99,9 @@ void Repeater::Stop() noexcept {
   if (!IsPlaying()) {
     return;
   }
-  performer_.Stop();
-  performer_.SetPosition(0.0);
-  if (instrument_.has_value()) {
+  performer_->Stop();
+  performer_->SetPosition(0.0);
+  if (instrument_ != nullptr) {
     instrument_->SetAllNotesOff();
   }
   index_ = -1;
@@ -213,4 +127,4 @@ bool Repeater::Update() noexcept {
   return true;
 }
 
-}  // namespace barely
+}  // namespace barely::internal
