+++ examples/demo/musician_demo.cpp
@@ -12,6 +12,7 @@
 #include <utility>
 #include <vector>
 
+#include "barelymusician/barelycomposer.h"
 #include "barelymusician/barelymusician.h"
 #include "barelymusician/common/random.h"
 #include "barelymusician/composition/duration.h"
