+++ examples/demo/repeater_demo.cpp
@@ -6,8 +6,8 @@
 #include <optional>
 #include <thread>
 
+#include "barelymusician/barelycomposer.h"
 #include "barelymusician/barelymusician.h"
-#include "barelymusician/components/repeater.h"
 #include "examples/common/audio_clock.h"
 #include "examples/common/audio_output.h"
 #include "examples/common/console_log.h"
