+++ platforms/unity/CMakeLists.txt
@@ -8,9 +8,8 @@ target_compile_definitions(
 )
 target_link_libraries(
   barelymusicianunity PUBLIC
+  barelycomposer
   barelymusician
-  barelymusician_components_arpeggiator
-  barelymusician_components_repeater
 )
 if(MSVC)
   target_sources(
