+++ platforms/unity/unity_android.lds
@@ -1,5 +1,32 @@
 VERS_1.0 {
   global:
+    /* barelycomposer */
+    BarelyArpeggiator_Create;
+    BarelyArpeggiator_Destroy;
+    BarelyArpeggiator_IsNoteOn;
+    BarelyArpeggiator_IsPlaying;
+    BarelyArpeggiator_SetAllNotesOff;
+    BarelyArpeggiator_SetGateRatio;
+    BarelyArpeggiator_SetInstrument;
+    BarelyArpeggiator_SetNoteOff;
+    BarelyArpeggiator_SetNoteOn;
+    BarelyArpeggiator_SetRate;
+    BarelyArpeggiator_SetStyle;
+    BarelyRepeater_Clear;
+    BarelyRepeater_Create;
+    BarelyRepeater_Destroy;
+    BarelyRepeater_IsPlaying;
+    BarelyRepeater_Pop;
+    BarelyRepeater_Push;
+    BarelyRepeater_PushSilence;
+    BarelyRepeater_SetInstrument;
+    BarelyRepeater_SetRate;
+    BarelyRepeater_SetStyle;
+    BarelyRepeater_Start;
+    BarelyRepeater_Stop;
+    BarelyScale_GetPitch;
+
+    /* barelymusician */
     BarelyInstrument_GetControl;
     BarelyInstrument_GetNoteControl;
     BarelyInstrument_IsNoteOn;
@@ -39,34 +66,8 @@ VERS_1.0 {
     BarelyPerformer_SetPosition;
     BarelyPerformer_Start;
     BarelyPerformer_Stop;
-    BarelyScale_GetPitch;
     BarelyTask_GetPosition;
     BarelyTask_SetPosition;
-
-    BarelyArpeggiator_Create;
-    BarelyArpeggiator_Destroy;
-    BarelyArpeggiator_IsNoteOn;
-    BarelyArpeggiator_IsPlaying;
-    BarelyArpeggiator_SetAllNotesOff;
-    BarelyArpeggiator_SetGateRatio;
-    BarelyArpeggiator_SetInstrument;
-    BarelyArpeggiator_SetNoteOff;
-    BarelyArpeggiator_SetNoteOn;
-    BarelyArpeggiator_SetRate;
-    BarelyArpeggiator_SetStyle;
-    BarelyRepeater_Clear;
-    BarelyRepeater_Create;
-    BarelyRepeater_Destroy;
-    BarelyRepeater_IsPlaying;
-    BarelyRepeater_Pop;
-    BarelyRepeater_Push;
-    BarelyRepeater_PushSilence;
-    BarelyRepeater_SetInstrument;
-    BarelyRepeater_SetRate;
-    BarelyRepeater_SetStyle;
-    BarelyRepeater_Start;
-    BarelyRepeater_Stop;
-
   local:
     *;
 };
